connect to multiplayer
players join a lobby and wait until the lobby capacity is reached. once the final player joins the lobby and the capacity is reached, all players in the lobby get taken to the multiplayer scene. each lobby's player count and each player's temporary nickname is now displayed in the console.