multiplayer scoreboard
enhanced multiplayer scene by importing assets from singleplayer scene. created some new scripts for multiplayer scene to integrate smoothly and to interact with the multiplayer scoreboard instead of the score in the singleplayer scene's assets. right now the score in the multiplayer scene is being kept in one of the players' scores.